refactor(frontend): the three 360 pages share one kit instead of importing a company screen
Deal360, Company360 and Person360 render the same parts — grounded prose with
citations, a section shell that tells "nothing here" apart from "you may not
read this", wire values turned into words. Those parts already existed and
already had multiple callers; what they did not have was a home.

They lived in company360.tsx, so person360.tsx and network.tsx imported
dealRoleLabel from a screen about companies, and dealstatus.tsx imported
SentenceList from one. Eighteen modules read from that file, and it had grown
to 3140 lines — six times the repo's 500-line cap.

This moves the entity-agnostic half to frontend/src/screens/record360/ and
points its twelve real consumers there. company360.tsx keeps everything that
is genuinely a company's and drops 506 lines.

The contract already agreed with the split: OrganizationBriefSentence is the
sentence type for the deal status card, Person360 and the growth-fit panel as
well as the org brief. Only its name says "Organization". The kit types
against it directly and calls it BriefSentence, so nobody reading record360 is
told a deal's sentence belongs to a company.

Behaviour is unchanged and no component was rewritten in the move. One
pre-existing slip is fixed on the way past: PeopleCard's doc comment had come
adrift onto incompleteGraph, and now sits back on PeopleCard.

The stylesheets keep their `co-` class prefix. Renaming those reaches four
files and every page that overrides one, which would bury this move in the
diff; it is a rename to do on its own.
