+++ frontend/src/screens/companygrowthfit.tsx
@@ -11,7 +11,7 @@ import { formatDateTime } from "../format/format";
 import { useLocale, useT } from "../i18n";
 import type { MessageKey } from "../i18n/en";
 import { problemMessageOf, throwProblem } from "./common";
-import { type BriefSentence, SentenceList, WrittenBy } from "./company360";
+import { type BriefSentence, SentenceList, WrittenBy } from "./record360";
 
 type GrowthFit = components["schemas"]["OrganizationGrowthFit"];
 type Band = GrowthFit["band"];
