+++ frontend/src/screens/record360/README.md
@@ -1,0 +1,40 @@
+# record360 — the shared kit behind Deal360, Company360 and Person360
+
+Three record pages answer the same question about different records: *where do
+we stand with this, and what do I do about it?* They render the same parts —
+grounded prose with citations, a section shell that knows the difference
+between "nothing here" and "you may not read this", wire values turned into
+words. This directory is those parts, owned in one place.
+
+It exists because the parts were already shared and had no home. They lived in
+`company360.tsx`, which grew to 3140 lines — six times the repo's 500-line file
+cap — and eighteen other modules imported from it. `person360.tsx` and
+`network.tsx` reached into a *company* screen for `dealRoleLabel`;
+`dealstatus.tsx` reached in for `SentenceList`. Every one of those imports was
+a shared component wearing one entity's name.
+
+## What belongs here
+
+A part belongs in the kit when it holds no opinion about WHICH record it is
+rendering. `SentenceList` renders sentences and their citations; it never asks
+whether they came from a deal or a company. That is the test.
+
+A part does NOT belong here because two pages happen to both use it today. A
+company's commercial panel and a deal's offer table are different components
+that look alike, and merging them produces one component with two modes and a
+flag — which is the thing this kit exists to avoid, not an instance of it.
+
+## The contract already agrees
+
+`OrganizationBriefSentence` is the sentence type for the org brief, the deal
+status card, Person360 and the growth-fit panel alike. Only its NAME says
+"Organization". The kit types against it directly and calls it what it is, so a
+reader of `record360` is not told that a deal's sentence is a company's.
+
+## The CSS keeps its `co-` prefix, for now
+
+The stylesheets still spell these classes `co-brief-lines`, `co-card` and so
+on, and this kit leaves them alone. Renaming them reaches four stylesheets and
+every page that overrides one, which would bury the extraction it was mixed
+into. The names are wrong and the styles are shared; that is a rename to do on
+its own.
